LayoutTests/fast/frames/seamless/seamless-window-location-replace-expected.txt

PASS frames[0].frames[0].document.body.textContent is "DONE\n"
